requirements.txt
python-telegram-bot
pandas
mplfinance
google-generativeai
pillow
requests
